Make UILayer and Container use shared_ptr
Originally UILayer and Container used and/or returned raw pointers to
objects. Although the memory was correctly deleted at the end of the
objects lifetimes, it can cause ambiguity on the users side on whether
or not they have to worry about the memory management. Changing to smart
pointer makes it obvious that no management is required on their part