Reset renderer projection before rendering
UILayer now resets the Renderer2D projection to be the default
position/resolution. Before, if a scene changed the camera position,
that change would persist through the UI rendering, causing the UI
widgets to be rendered at an incorrect position. The hitbox of the
widgets would not change, making it seem like the widgets are
unresponsive. This fixes that bug