Update AddWidget function
Changed AddWidget function to take in a variable amount of arguments,
instead of a pointer to a widget. It now acts the same way as
make_unique<T> would. This removes the redundancy of having to specify
the widget type twice. i.e "AddWidget<Button>(new Button)" and creates
cleaner code.