Create SubTexture class and update accordingly
This is the second and last step to fixing the bug where a Subtexture's
source texture would get destroyed before the lifetime of the subtexture
resulting in corrupted images. Requiring Textures to be acquired as
shared_ptrs and having a SubTexture class hold a shared_ptr to it's
source should stop this from occuring.