Update Sprite to take place of SubTexture class
It was getting unwieldy to have both a texture and subtexture class. In
addition, adding scale, size, rotation, etc to each of those would
result in a lot of duplicated code. The solution was to make Texture
just about the pixel, size and opengl data, and to use Sprite as a
representation of a texture, with its own size, texture coordinates,
scale, color, etc.