Improved Audio Player
Audio Player now is in charge of everything sound related.
SoundEffects/AudioStreams no longer exist. Instead the when you load a
sound file, you can ask the audioplayer to stream it, and it will
automatically take care of streaming. This simplifies the API and should
make it easier to expand in the future.