Fix: Fix vertical position alignment of text
- Change how GetStringPixelSize() calculates height
- Change tallest_glyph_height to highest_glyph_bearing to be more
  accurate

In order to allow text position to be specified by the top-left corner
and all text rendered at the same height to have the same baseline, the
vertical size of text in any one font is exactly the same. Unless a '|'
character is used, there will almost always be some padding between the
size of the text and where the actual pixels of each glyphs start. This
causes text to look misplaced when trying to center vertically. In order
to fix this, the padding is doubled to distribute that difference equally.
This does increase the amount of empty vertical space, but fixes the
improper alignment of the text.